requirements.txt
# Shared
opencv-python
numpy
Pillow
tqdm
kaggle
matplotlib

# Diffusion Model (PyTorch)
torch
torchvision
diffusers[torch]
accelerate
transformers

# GAN & VAE (TensorFlow)
tensorflow
